Use vfsStream in testing ScanCommand
vfsStream allows the creation of a virtual filesystem that exists only
in memory, meaning temporary files are truly temporary and don't rely on
an actual, live filesystem.

The only problem here is that SplFileInfo::getRealPath() won't work with
it, so I had to change File. I don't see a reason getPathname() won't
work in this instance.