Test for dots if the output stream isn't decorated
The code was correct, but the unit tests weren't catching if the check
for a decorated output stream was missing.

This was discovered using humbug (config and .gitignore update
included). Note that, for now, tests/Subscriber/XmlTest.php breaks
humbug (due to it's creation of files), so it should be temporarily
removed when running humbug.